# docs/swagger.yaml
basePath: /v1
definitions:
  models.Segment:
    properties:
      ID:
        type: integer
      Slug:
        type: string
    type: object
  models.User:
    properties:
      ID:
        type: integer
    type: object
  v1.response:
    properties:
      error:
        example: message
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a service for dynamic user segmentation.
  title: DynamicUserSegmentationService
  version: "1.0"
paths:
  /v1/:
    get:
      consumes:
      - application/json
      description: Get Active Segments By ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Segment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get Active Segments By ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Change User`s Active Segments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Change Segments
      tags:
      - users
  /v1/segment/:
    delete:
      consumes:
      - application/json
      description: Delete segment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Segment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Delete segment
      tags:
      - segments
    post:
      consumes:
      - application/json
      description: Create segment
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Segment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Create segment
      tags:
      - segments
swagger: "2.0"